Fix compilation error: change int to double for exposure times in RotationManager
- Updated exposure variable declarations from int to double
- Fixes incompatible types error with PPMPreferences.get*ExposureMs() return values
- Maintains consistency with AngleExposure record expecting double exposureMs

+++ src/main/java/qupath/ext/qpsc/modality/ppm/RotationManager.java
@@ -67,10 +67,10 @@ public class RotationManager {
             }
 
             // Get exposure times from PPMPreferences
-            int plusExposure = PPMPreferences.getPlusExposureMs();
-            int minusExposure = PPMPreferences.getMinusExposureMs();
-            int zeroExposure = PPMPreferences.getZeroExposureMs();
-            int uncrossedExposure = PPMPreferences.getUncrossedExposureMs();
+            double plusExposure = PPMPreferences.getPlusExposureMs();
+            double minusExposure = PPMPreferences.getMinusExposureMs();
+            double zeroExposure = PPMPreferences.getZeroExposureMs();
+            double uncrossedExposure = PPMPreferences.getUncrossedExposureMs();
 
             strategies.add(new PPMRotationStrategy(
                     new AngleExposure(plusTick, plusExposure),
